backend_django/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.2.12
django-redis==5.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.4.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
wrapt==1.14.1
